refactor: ♻️ Extract script loader to new file
