+++ modules/ppcp-sdk-v6/resources/js/sdkLoader.js
@@ -5,8 +5,8 @@
  */
 
 import { postJson } from './utils/api';
+import { loadScript } from './utils/scriptLoaders';
 
-const scriptPromises = {};
 let instancePromise = null;
 
 const PAGE_TYPE_MAP = {
@@ -15,35 +15,6 @@ const PAGE_TYPE_MAP = {
 	checkout: 'checkout',
 	'mini-cart': 'mini-cart',
 };
-
-/**
- * Dynamically loads the PayPal SDK v6 core script.
- *
- * The load promise is cached per URL (not sniffed from the DOM) so a
- * failed load rejects every awaiting caller and clears the cache,
- * allowing a later retry to insert a fresh script tag.
- *
- * @param {string} sdkUrl - The SDK URL.
- * @return {Promise<void>} Resolves when the script is loaded.
- */
-function loadSdkScript( sdkUrl ) {
-	if ( ! scriptPromises[ sdkUrl ] ) {
-		scriptPromises[ sdkUrl ] = new Promise( ( resolve, reject ) => {
-			const script = document.createElement( 'script' );
-			script.src = sdkUrl;
-			script.async = true;
-			script.onload = resolve;
-			script.onerror = () => {
-				script.remove();
-				delete scriptPromises[ sdkUrl ];
-				reject( new Error( 'Failed to load PayPal SDK v6 script.' ) );
-			};
-			document.head.appendChild( script );
-		} );
-	}
-
-	return scriptPromises[ sdkUrl ];
-}
 
 /**
  * Loads the SDK script, fetches a client token and creates the instance.
@@ -78,7 +49,7 @@ export function loadSdkV6( config, context ) {
  */
 async function createInstance( config, context ) {
 	const [ , tokenData ] = await Promise.all( [
-		loadSdkScript( config.sdk_url ),
+		loadScript( config.sdk_url ),
 		postJson( config.ajax.client_token ),
 	] );
